refactor: enviando Props e Type para arquivo externo
